+++ src/dodal/devices/b07/__init__.py
@@ -1,0 +1,3 @@
+from dodal.devices.b07.grating import B07BGrating
+
+__all__ = ["B07BGrating"]
