+++ src/dodal/devices/b07_1/__init__.py
@@ -1,0 +1,3 @@
+from dodal.devices.b07_1.grating import B07CGrating
+
+__all__ = ["B07CGrating"]
