+++ src/dodal/devices/i09/__init__.py
@@ -1,0 +1,4 @@
+from dodal.devices.i09.dcm import DCM
+from dodal.devices.i09.grating import I09Grating
+
+__all__ = ["DCM", "I09Grating"]
