+++ src/dodal/devices/p60/__init__.py
@@ -1,0 +1,3 @@
+from dodal.devices.p60.lab_xray_source import LabXraySource, LabXraySourceReadable
+
+__all__ = ["LabXraySource", "LabXraySourceReadable"]
